feat: add trading panel and positions info to flatcoin mobile layout
